==> 1-LiteratureHandler.ipynb <==
# Literature Processer
* Jacob Yousif

## Importing the libraries

import warnings
warnings.filterwarnings('ignore')

%%capture
!pip install chardet

import os
import re
import chardet
import pandas as pd
import json
from sklearn.preprocessing import LabelEncoder

## The path for the source of the literature

directory = 'Literature/'

## Predefined functions

def extract_file(file_path):
    
    with open(file_path, 'r') as file:
        data = json.load(file)
    return data

def count_words(text):
    if pd.isna(text):
        return 0
    return len(text.split())

def remove_whitespace(text):
    return  " ".join(text.split())

def filename_to_title(filename):
    name_without_ext = filename.replace(".txt", "")
    title = " ".join(
        [
            word if i == 0 else word.capitalize()
            for i, word in enumerate(re.findall("[A-Z][^A-Z]*", name_without_ext))
        ]
    )
    return title

def read_text_file(file_path):
    with open(file_path, 'rb') as file:
        raw_data = file.read()
    result = chardet.detect(raw_data)
    encoding = result['encoding']
    with open(file_path, 'r', encoding=encoding) as file:
        text = file.read()
    
    return text

## The processing

books_info = extract_file('Datasets/books.json')

print(books_info)

files = [f for f in os.listdir(directory) if os.path.isfile(os.path.join(directory, f))]

files.sort()

df = pd.DataFrame()

df['Book'] = [filename_to_title(book) for book in files]

books = [read_text_file(directory + item) for item in files]

df['Text'] = books

books_info = extract_file('Datasets/books.json')

df['Author'] = [books_info[x] for x in files]

df['Text'] = df['Text'].apply(remove_whitespace)

df['Size'] = df['Text'].apply(count_words)

_sum = df['Size'].sum()

print(f"Sum of all words in all books: {_sum} words.")

label_encoder = LabelEncoder()
df['AuthorCode'] = label_encoder.fit_transform(df['Author'])

unique_classes = df['Author'].unique()
num_classes = len(unique_classes)

print(f"Unique Authors: {unique_classes}\n")
print(f"Number of unique classes: {num_classes}")

author_dict = df.set_index('Author')['AuthorCode'].to_dict()

with open('Datasets/author_codes.json', 'w') as json_file:
    json.dump(author_dict, json_file)

author_dict = dict(sorted(author_dict.items(), key=lambda item: item[1]))

author_dict

has_nan = df.isna().any().any()
print(f"DataFrame has NaN values: {has_nan}")

df['Book']

df

csv_file_path = 'Datasets/Literature.csv'
df.to_csv(csv_file_path, index=False)
