==> 2-TextSegmenter.ipynb <==
# Text Segmmenter

*   Jacob Yousif


## Importing the libraries

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
from nltk.tokenize import sent_tokenize

## Loading the data

file_path = 'Datasets/Literature.csv'
df = pd.read_csv(file_path)

has_nan = df.isna().any().any()
print(f"DataFrame has NaN values: {has_nan}")

## Predefined functions

def create_segments(sequence, segment_size=15, buffer=5):
    sentences = sent_tokenize(sequence)
    num_of_chunks = ((len(sentences) - segment_size) // buffer) + 1
    segments = [" ".join(sentences[i:i + segment_size]) for i in range(0, num_of_chunks * buffer, buffer)]
    return segments

## The segmentation

columns = ['Book', 'Text', 'Author', 'AuthorCode']
result = pd.DataFrame(columns=columns)

for index, row in df.iterrows():
    segments = create_segments(row['Text'])
    for segment in segments:
        _row = {
        'Book': row['Book'],
        'Text': segment,
        'Author': row['Author'],
        'AuthorCode': row['AuthorCode']
        }
        result.loc[len(result)] = _row

result

print('The size of the dataset is:', {len(result)})

has_nan = result.isna().any().any()
print(f"DataFrame has NaN values: {has_nan}")

csv_file_path = 'Datasets/SegmentedLiterature.csv'
result.to_csv(csv_file_path, index=False)
