==> 5-DatasetDivider.ipynb <==
# Dataset Divider

*   Jacob Yousif

## Importing the libraries

import warnings
warnings.filterwarnings('ignore')

%%capture
!pip install imbalanced-learn

%%capture
import numpy as np
import pandas as pd
from sklearn.preprocessing import LabelEncoder, StandardScaler
from sklearn.model_selection import train_test_split
from imblearn.over_sampling import SMOTE

## Loading the data

file_path = 'Datasets/ProcessedSegmentedLiterature.csv'
df = pd.read_csv(file_path)
df= df.sort_values(by='AuthorCode', ascending=True)

## The processing

x_columns = [ 
    "AverageWordLength", "AverageSentenceLength", "AverageSyllablePerWord",
    "PunctuationCount", "FunctionalWordsCount", "TypeTokenRatio", "HonoreMeasureR",
    "Hapax", "SichelesMeasureS", "Dihapax", "YulesCharacteristicK", "SimpsonsIndex",
    "BrunetsMeasureW", "ShannonEntropy", "FleschReadingEase", "FleschKincaidGradeLevel",
    "DaleChallReadability", "GunningFog"
]
y_column = ['AuthorCode']

X = df[x_columns].values

y = df[y_column].values

smote = SMOTE()
X, y = smote.fit_resample(X, y)

from collections import Counter
class_distribution = Counter(y)
print("Class Distribution:")
for cls, count in class_distribution.items():
    print(f"Class {cls}: {count}")

print('The size of the dataset is:', len(X))

X_train, X_temp, y_train, y_temp = train_test_split(X, y, test_size=0.3, stratify=y, random_state=42)  
X_val, X_test, y_val, y_test = train_test_split(X_temp, y_temp, test_size=0.5, stratify=y_temp, random_state=42)

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_val = scaler.transform(X_val)
X_test = scaler.transform(X_test)

print('The training set (Descriptive Features) size:', len(X_train))

print('The training set (Target Feature) size:', len(y_train))

print('The validation set (Descriptive Features) size:', len(X_val))

print('The validation set (Target Feature) size:', len(y_val))

print('The test set (Descriptive Features) size:', len(X_test))

print('The test set (Target Feature) size:', len(y_test))

np.savez('Datasets/LearningSet.npz', 
         X_train=X_train, y_train=y_train, 
         X_val=X_val, y_val=y_val, 
         X_test=X_test, y_test=y_test)
