==> 6-XGBoostHypertuner.ipynb <==
# XGBoost Hypertuning

*   Jacob Yousif

## Importing the libraries

import warnings
warnings.filterwarnings('ignore')

%%capture
!pip install imbalanced-learn xgboost optuna

%%capture

import pandas as pd
import numpy as np
import joblib
from sklearn.metrics import accuracy_score
from xgboost import XGBClassifier
import optuna
import time

## Loading the data

data = np.load('Datasets/LearningSet.npz')
X_train = data['X_train']
y_train = data['y_train']
X_val = data['X_val']
y_val = data['y_val']
X_test = data['X_test']
y_test = data['y_test']

## Tuning the model

results = []

total = 0

def objective(trial):
    global total
    global results
    
    params = {
        'n_estimators': trial.suggest_int('n_estimators', 20, 500),
        'max_depth': trial.suggest_int('max_depth', 3, 25),
        'num_leaves': trial.suggest_int('num_leaves', 10, 150),
        'learning_rate': trial.suggest_float('learning_rate', 0.01, 0.5),
        'min_child_samples': trial.suggest_int('min_child_samples', 10, 200),
        'subsample': trial.suggest_float('subsample', 0.1, 1.0),
        'colsample_bytree': trial.suggest_float('colsample_bytree', 0.5, 1.0),
        'gamma': trial.suggest_float('gamma', 0.0, 5.0),
        'lambda': trial.suggest_float('lambda', 0.0, 5.0),
        'alpha': trial.suggest_float('alpha', 0.0, 5.0),
        'scale_pos_weight': trial.suggest_float('scale_pos_weight', 0.5, 5.0),
    }
    
    start_time = time.time()
    clf = XGBClassifier(**params, use_label_encoder=False, eval_metric='mlogloss')
    clf.fit(X_train, y_train, eval_set=[(X_val, y_val)], verbose=False, early_stopping_rounds=3)
    y_pred = clf.predict(X_val)
    accuracy = accuracy_score(y_val, y_pred)
    duration = time.time() - start_time

    total += duration
    
    results.append({
        'Trial': trial.number,
        'Accuracy': accuracy,
        'Duration': duration,
        'Params': params,
        'BestAccuracy': max([res['Accuracy'] for res in results], default=0)
    })
    
    return accuracy

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=500)

seconds = int(total)
print('In seconds, the total duration of the hypertuning process is:', seconds, 'seconds.')

minutes = int(total / 60)
print('In minutes, the total duration of the hypertuning process is:', minutes, 'minutes.')

hours = round(total / 3600, 1)
print('In hours, the total duration of the hypertuning process is:', hours, 'hours.')

progress = pd.DataFrame(results)

progress

csv_file_path = 'Datasets/XGBoostHypertuningProcess.csv'
progress.to_csv(csv_file_path, index=False)

best_params = study.best_params
print("Best Parameters:", best_params)

clf = XGBClassifier(**best_params, use_label_encoder=False, eval_metric='mlogloss')

clf.fit(X_train, y_train)

joblib.dump(clf, 'Model/TrainedXGBoostModel.joblib')
